Master Restructure Merge Fixes
