+++ client/src/App.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
-import Header from './components/Header';
-import Main from './components/Main';
+import { Main, Header } from './components/Main';
 
 /*
 Component that renders our Header and Main component.
